+++ packages/ui/exports/overlays-test-suites.js
@@ -1,1 +1,2 @@
 export { runOverlayMixinSuite } from '../components/overlays/test-suites/OverlayMixin.suite.js';
+export { runWithHoverInteractionSuite } from '../components/overlays/test-suites/WithHoverInteraction.suite.js';